requirements.txt
opencv-python==4.4.0.46
numpy==1.18.5
imutils==0.5.3
Pillow==8.0.1
streamlit==0.71.0
opencv-contrib-python==4.4.0.46
opencv-python-headless==4.2.0.32
